Adds sensor for every Betriebsmodus and Total kWh counters
Betriebsmodus is a bit coded 15 bit integer. Each bit represents
a different symbol on the WP. This can also be visualized now using
the picture entity card. Currently not all images are supported.

Sensors for Wh and kWh for warm water and heating are combined and
should increase monotonically over the day and be reset at midnight
to allow visualisation in energy dashboard