Fixes RAUMISTTEMP
Prior to this commit it was not possible to set the RAUMISTTEMP.
It was always reset by the WP to the RAUMSOLLTEMP.

You can now use a temperature entity from Home Assistant and set
RAUMISTTEMP with it. This renders FEK useless, cause the same can
be achieved via HA <3. Same should be possible for FEUCHTE.