Adds proper mode handling to climates
Apparently the heating switches to cooling when SOMMERBETRIEB
is active. Means that suddenly it looks like WP is set to heat
up to 25° instead of cooling down to 25°, at least in my case.
This is very counter intuitive, because the mode switch is
not visible in HA. This is the first try to see, if proper
mode handling improves the situation, or if there are more
checks needed.