Fix broken fan control
After reboot the fans were turned off. This was happening due to
a missing config option:

restore_mode -
ALWAYS_OFF (Default) - Always initialize the fan as OFF on bootup.

Fixing this revealed another bug in the code and in ESPHOME.
A workaround is in place until the issue is fixed. Fan state and
fan speed have to be sent in two consecutive calls.

Fixes: #18