Fix fan turn_on()
Turn_on() cannot be left blank. Doing so will not send a message,
if the fan was only turned on and no speed was set. First time
the fan is turned on, the speed will be set to 100%. All other
calls should reset the speed to the value that was set before
it had been turned off.

Fixes: #18