introduces variant handling
In order to support settings for different heat pump variants,
there is now a base config that defines the common parts and a
specific yaml for every variant. Specific parts can be moved out
of the base and defined for each heat pump. If there are changes
required on C++ level, C macros can be used to change the source
code dynamically.

    // do this
    // do that
    // do the default