Update energy consumption more frequently
With a period of 1 hour it is very likely that a huge chunk
of kWh is mapped to the wrong hour in the energy dashboard
of home assistant. This became obvious with the latest update
that shows untracked energy consumption. First we miss to count
then we count too much which looks ugly.