Shift daily energy combined sensors
It might not be a good idea to directly query the new daily value
directly some seconds after midnight. If the time of the heatpump
does not exactly match the sntp time this can lead to wrong
behavior. So we decouple it a bit and in the next step also sync
the clock automatically.

Fixes: #85