Hot water needs to be split into two entities
https://github.com/kr0ner/OneESP32ToRuleThemAll/issues/50
already revealed that setting only one solltemp can lead
to toggling behavior. Should've fixed it for hot water
as well. If a schedule for hot water is active (not recommended)
in the heat pump, there are two climate controllers needed to
change the values, depending on the time of day.

Also fixed the namings since there should always be *SOLL_TAG &
*SOLL_NACHT