Handle external events properly
The current implementation relies on polling the values
in order to keep the sensors up to date. If the heat pump is
being controlled via the display, it takes up to <interval_time>
to reflect the change in HA user interface. For this to work we
nee to add another callback that handles this use case, but only
for values that can be actively changed. The second use case is
when updating settings from within HA. Also here we'd require
some time until the values are polled and updated again. We can
improve the situation by querying the sensor value everytime an
update request is sent to the heatpump.